Backport from main: Fix: #4989 - mismatch between individual and family lists
