+++ app/Module/FamilyListModule.php
@@ -19,52 +19,35 @@ declare(strict_types=1);
 
 namespace Fisharebest\Webtrees\Module;
 
-use Fig\Http\Message\StatusCodeInterface;
-use Fisharebest\Webtrees\Auth;
 use Fisharebest\Webtrees\I18N;
-use Fisharebest\Webtrees\Registry;
-use Fisharebest\Webtrees\Validator;
-use Psr\Http\Message\ResponseInterface;
-use Psr\Http\Message\ServerRequestInterface;
 
 /**
  * Class FamilyListModule
  */
-class FamilyListModule extends IndividualListModule
+class FamilyListModule extends AbstractIndividualListModule
 {
-    protected const ROUTE_URL = '/tree/{tree}/family-list';
+    protected function routeUrl(): string
+    {
+        return '/tree/{tree}/family-list';
+    }
 
-    // The individual list and family list use the same code/logic.
-    // They just display different lists.
-    protected bool $families = true;
+    protected function showFamilies(): bool
+    {
+        return true;
+    }
 
-    /**
-     * How should this module be identified in the control panel, etc.?
-     *
-     * @return string
-     */
     public function title(): string
     {
         /* I18N: Name of a module/list */
         return I18N::translate('Families');
     }
 
-    /**
-     * A sentence describing what this module does.
-     *
-     * @return string
-     */
     public function description(): string
     {
         /* I18N: Description of the “Families” module */
         return I18N::translate('A list of families.');
     }
 
-    /**
-     * CSS class for the URL.
-     *
-     * @return string
-     */
     public function listMenuClass(): string
     {
         return 'menu-list-fam';
